+++ src/plugins/naive.ts
@@ -1,9 +1,9 @@
 import type { App } from 'vue';
 import {
   create,
-  NConfigProvider,
   NMessageProvider,
   NDialogProvider,
+  NConfigProvider,
   NInput,
   NButton,
   NForm,
@@ -66,8 +66,10 @@ import {
   NTimePicker,
   NBackTop,
   NSkeleton,
+  NCascader,
 } from 'naive-ui';
 
+// https://www.naiveui.com/en-US/os-theme/docs/import-on-demand
 const naive = create({
   components: [
     NMessageProvider,
@@ -135,6 +137,7 @@ const naive = create({
     NTimePicker,
     NBackTop,
     NSkeleton,
+    NCascader,
   ],
 });