+++ src/views/comp/table/basic.vue
@@ -10,11 +10,7 @@
       :actionColumn="actionColumn"
       :scroll-x="1360"
       @update:checked-row-keys="onCheckedRow"
-    >
-      <template #toolbar>
-        <n-button type="primary" @click="reloadTable">刷新数据</n-button>
-      </template>
-    </BasicTable>
+    />
   </n-card>
 </template>
 
@@ -24,6 +20,7 @@
   import { getTableList } from '@/api/table/list';
   import { columns } from './basicColumns';
   import { useDialog, useMessage } from 'naive-ui';
+  import { DeleteOutlined, EditOutlined } from '@vicons/antd';
 
   const message = useMessage();
   const dialog = useDialog();
@@ -31,17 +28,17 @@
 
   const params = reactive({
     pageSize: 5,
-    name: 'xiaoMa',
+    name: 'NaiveAdmin',
   });
 
   const actionColumn = reactive({
-    width: 150,
+    width: 180,
     title: '操作',
     key: 'action',
     fixed: 'right',
     align: 'center',
     render(record) {
-      return h(TableAction, {
+      return h(TableAction as any, {
         style: 'button',
         actions: createActions(record),
       });
@@ -52,21 +49,16 @@
     return [
       {
         label: '删除',
-        icon: 'ic:outline-delete-outline',
+        // 配置 color 会覆盖 type
+        icon: DeleteOutlined,
         onClick: handleDelete.bind(null, record),
-        // 根据业务控制是否显示 isShow 和 auth 是并且关系
-        ifShow: () => {
-          return true;
-        },
         // 根据权限控制是否显示: 有权限，会显示，支持多个
         auth: ['basic_list'],
       },
       {
         label: '编辑',
+        icon: EditOutlined,
         onClick: handleEdit.bind(null, record),
-        ifShow: () => {
-          return true;
-        },
         auth: ['basic_list'],
       },
     ];
@@ -78,10 +70,6 @@
 
   function onCheckedRow(rowKeys) {
     console.log(rowKeys);
-  }
-
-  function reloadTable() {
-    actionRef.value.reload();
   }
 
   function handleDelete(record) {