+++ src/views/comp/table/editCell.vue
@@ -11,11 +11,7 @@
       @edit-change="onEditChange"
       @update:checked-row-keys="onCheckedRow"
       :scroll-x="1360"
-    >
-      <template #toolbar>
-        <n-button type="primary" @click="reloadTable">刷新数据</n-button>
-      </template>
-    </BasicTable>
+    />
   </n-card>
 </template>
 
@@ -28,7 +24,7 @@
   const actionRef = ref();
   const params = reactive({
     pageSize: 5,
-    name: 'xiaoMa',
+    name: 'NaiveAdmin',
   });
 
   function onEditChange({ column, value, record }) {
@@ -46,12 +42,7 @@
     console.log(rowKeys);
   }
 
-  function reloadTable() {
-    console.log(actionRef.value);
-    actionRef.value.reload();
-  }
-
-  function editEnd({ record, index, key, value }) {
+  function editEnd({ value }) {
     console.log(value);
   }
 </script>