+++ src/views/comp/table/editRow.vue
@@ -12,11 +12,7 @@
       @edit-change="onEditChange"
       @update:checked-row-keys="onCheckedRow"
       :scroll-x="1590"
-    >
-      <template #toolbar>
-        <n-button type="primary" @click="reloadTable">刷新数据</n-button>
-      </template>
-    </BasicTable>
+    />
   </n-card>
 </template>
 
@@ -30,7 +26,7 @@
   const currentEditKeyRef = ref('');
   const params = reactive({
     pageSize: 5,
-    name: 'xiaoMa',
+    name: 'NaiveAdmin',
   });
 
   const actionColumn = reactive({
@@ -101,12 +97,7 @@
     console.log(rowKeys);
   }
 
-  function reloadTable() {
-    console.log(actionRef.value);
-    actionRef.value.reload();
-  }
-
-  function editEnd({ record, index, key, value }) {
+  function editEnd({ value }) {
     console.log(value);
   }
 </script>