+++ src/views/comp/upload/index.vue
@@ -7,7 +7,7 @@
       <n-grid cols="2 s:1 m:3 l:3 xl:3 2xl:3" responsive="screen">
         <n-grid-item offset="0 s:0 m:1 l:1 xl:1 2xl:1">
           <n-form
-            :label-width="80"
+            :label-width="90"
             :model="formValue"
             :rules="rules"
             label-placement="left"
@@ -29,7 +29,7 @@
                 name="files"
                 :width="100"
                 :height="100"
-                @uploadChange="uploadChange"
+                @upload-change="uploadChange"
                 v-model:value="formValue.images"
                 helpText="单个文件不超过2MB，最多只能上传10个文件"
               />